weighted malis now works

+++ malis/malis_cpp.cpp
@@ -34,14 +34,14 @@ void malis_loss_weights_cpp(const int nVert, const int* seg,
 
 
     /* Disjoint sets and sparse overlap vectors */
-    vector<map<int,int> > overlap(nVert);
+    vector<map<int,uint64_t> > overlap(nVert);
     vector<int> rank(nVert);
     vector<int> parent(nVert);
     boost::disjoint_sets<int*, int*> dsets(&rank[0],&parent[0]);
     for (int i=0; i<nVert; ++i){
         dsets.make_set(i);
         if (0!=seg[i]) {
-            overlap[i].insert(pair<int,int>(seg[i],1));
+            overlap[i].insert(pair<int,uint64_t>(seg[i],1));
         }
     }
 
@@ -60,8 +60,8 @@ void malis_loss_weights_cpp(const int nVert, const int* seg,
     /* Start MST */
     int e;
     int set1, set2;
-    int nPair = 0;
-    map<int,int>::iterator it1, it2;
+    uint64_t nPair = 0;
+    map<int,uint64_t>::iterator it1, it2;
 
     /* Start Kruskal's */
     for (unsigned int i = 0; i < pqueue.size(); ++i ) {
@@ -97,7 +97,7 @@ void malis_loss_weights_cpp(const int nVert, const int* seg,
             while (it2 != overlap[set2].end()) {
                 it1 = overlap[set1].find(it2->first);
                 if (it1 == overlap[set1].end()) {
-                    overlap[set1].insert(pair<int,int>(it2->first,it2->second));
+                    overlap[set1].insert(pair<int,uint64_t>(it2->first,it2->second));
                 } else {
                     it1->second += it2->second;
                 }
